add administrator

+++ jcart-administration-back/src/main/java/io/cjf/jcartadministrationback/po/Order.java
@@ -21,7 +21,7 @@ public class Order {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
